Update atmat/atphysics/TouschekPiwinski/MomAperture_Projecect2Start.m
New function MomAperture_Project2Start:

Similar to MomAperture_allRing, but all positions are first tracked until the ring starting point and then tracked together as many turns as required. It speeds up the calculations around a 50%, results are similar but not exactly the same since the searching method is also different (bipartition).